Reverted to non auth.user authentication...
After deciding on a quicker and easier (albeit less secure) method we reverted to a non auth.user method of authentication.
This method involves using a bearer token to verify a user is authenticated. A new field is added to the model for user called token.

New methods Logout and authenticate have been added. They will serve to prove identity.